fix(dev): unmarshal error on Diff1Work as int64

+++ types.go
@@ -727,7 +727,7 @@ type Devs struct {
 	LastSharePool       int64   `json:"Last Share Pool"`
 	LashShareTime       int64   `json:"Lash Share Time"`
 	TotalMH             float64 `json:"TotalMH"`
-	Diff1Work           int64   `json:"Diff1 Work"`
+	Diff1Work           float64 `json:"Diff1 Work"`
 	DifficultyAccepted  float64 `json:"Difficulty Accepted"`
 	DifficultyRejected  float64 `json:"Difficulty Rejected"`
 	LastShareDifficulty float64 `json:"Last Share Difficulty"`
